Refactor visibility Document and upgrate schema verion to 1.
VisibilityDocument is the setting entity and store all visibility
information of a schema. It seems to be not efficient to have
multiple params like VisibilityToPackages and
VisibilityToSystem to be passed around and only convert to
VisibilityDocument at the end.

This change VisibilityDocument's schema and the logic of set and look
up Visibility setting.

Test: presubmit
Bug: 202194495
Change-Id: I73d2657891e3b163306c25ff2693874cbf57b02e