Add ParcelableDocuments to AppSearch.
The Parcelable object contains a List of {@link GenericDocument}.
It will write the data via Parcel.writeBlob.
If the blob is less than 16 KiB, then it is stored in-place, otherwise
it is transferred by way of an anonymous shared memory region.

System healthy test result analysis: http://shortn/_fWOEQvcWm2

Deleted testPutDocument_throwsNullException since it won't work with the
new put mechanism. It's trying to verify the Exception could be passed
correctly and this is already covered by
AppSearchDeviceTest.createSessionInStoppedUser in CTS hostsidetests.

Bug: 185441119
Test: presubmit
Change-Id: Ief1d185ffdc6f36d5c0dc638b921d407101d69d9