ContactsIndexer implementation (1)
Added 3 classes

- ContactsIndexerManagerService
- ContactsContentObserver
- ContactsIndexerPerUser

to handle the contact change notifications from CP2.

The overall structure and threading are added in this change, but may
require some adjustment in the follow-up CLs.

In the follow-up CLs, we would mainly implement the Delta Update to:
- retrieve contact updates from CP2.
- create AppSearch documents based on those updates, and index them into
AppSearch.

One Pager: go/appsearch-contacts-indexer-one-pager
Design doc: go/appsearch-contacts-indexer(Implementation Plan:
https://docs.google.com/document/d/1qfek20XQQYh02V2J6K0pXfrhnXYmVH-1NqwI8Xiqd10/edit#heading=h.8tlk1jy5d7ek)

Bug: 203605504
Test: manual test on a device with muli-user to verify the timestamps
are persisted to the disk whenever there is some contact change.

Change-Id: I7f3f9335e06fc5c03d5b37e5bfb98c9fcf4fe301