Move visibility Document to the SDK side.
VisibilityDocument is the setting entity and store all visibility
information of a schema. It seems to be not efficient to have
multiple params like VisibilityToPackages and
VisibilityToSystem to be passed around and only convert to
VisibilityDocument at the end.

This change move building VisibilityDocument to SDK side so that we
can pass it's bundle only via binder rather than 4 params
(visibilityForPermission and VisibilityForRole is coming).

Test: presubmit
Bug: 202194495
Change-Id: I99731f5bea43ba4e92840311b7cb06a8812ddad8