Add verification that target package must exist in target user.
AppSearch currently supports cross-user API requests if the caller
user has INTERACT_ACROSS_USERS_FULL.
This means that a package with this permission could interact
with AppSearch's data in some other target user, even if the
package isn't installed on the target user.

This change fix it and verify it's not allowed anymore.
This change all add a SystemApi to help testing this scenario.

Since this change will create a package context, here is the
SH test number : http://shortn/_doxtdqkPDf

Bug: 193423599
Bug: 193423447
Bug: 193423599
Test: atest --rebuild-module-info CtsAppSearchHostTestCases:AppSearchMultiUserTest
Change-Id: I8f5621a17bd615fa0d7c5f529e4a79375adf4d96