Rework how executors are managed in AppSearchManagerService.
* Executors are now created per each device user.
* When a user locks or stops, its executor is drained and removed.
* Some common functionality around user verification moves to a
  helper class.

This CL is intended to fix the issue where jobs remain in the AppSearch
executor queue after a user is locked, leading to exceptions from trying
to use a closed AppSearchImpl instance.

This situation is very hard to reliably reproduce in a test and I haven't
been able to do so, so this fix will have to be validated by letting it
run for a while in postsubmit.

Bug: 219640103
Bug: 218419277
Bug: 217173331
Test: Behaviour after a user locks tested with AppSearchMultiUserTest#testCreateSessionInStoppedUser
Test: Not breaking any existing functionality tested via all cts and presubmit tests
Change-Id: If58142f5284fd143cbc5326dc7aa01d8e647cc0c