Switch ordering of the user stopping calls.
With the current order, the executor is removed from the
executor map and then shut down. Any activity that occurs
during executor shutdown will re-create a new executor in
the map and schedule itself, then crash because the AppSearchImpl
instance gets closed.

With the altered order, we first mark the user locked to ensure
exceptions are delivered to future callers and no more calls are
scheduled onto the executor.

Crash clusters being monitored:
- http://shortn/_jEAwW9SbFX
- http://shortn/_96TsweaXS7

Bug: 218419277
Test: Will have to be tested via watching the crash clusters in postsubmit
Change-Id: Ieebd5bbffd92545b1c15d2c145744e5598207e95