Improve full update of contacts in AppSearch
Retrieve IDs of all contacts indexed in AppSearch, and diff them against
the IDs of all CP2 contacts to generate wanted (in CP2) and unwanted (in
AppSearch but not in CP2) IDs. Index the wanted contacts into the
builtin:Person corpus and remove the unwanted contacts from the same
corpus, resulting in a robust full update.

Schedule full update on bootstrap as well as when a configurable amount
of time has passed (default 30 days).

Bug: 203605504
Test: manually using acloud on specialist cloudtop; atest AppSearchServicesTests
Change-Id: I51bed1d7a9ae1528a531c46c39b1cde126debf43