Implement fingerprint to reduce unnecessary document updates
Add a new property fingerprint in Person schema. The fingerprint is a
MD5 hashcode of the String value of the GenericDocument, with fingerprint
array empty and creationTimestamp set to 0.

This property is not exposed in the Java class in Jetpack. But can be
accessed by using the GenericDocument directly.

During a delta update, for any updates to existing contacts, we try to
get the fingerprint of those first to see if we need to do the put.
Since right now we only track a smaller set of properties from CP2 for
the contacts, this would help us reduce the number of updates we make to
AppSearch.

For exsiting contacts in AppSearch, we won't update the
creationTimestamp.

Bug: 219730464
Test: AppSearchServicesTests
Change-Id: Ic54665de9fa869fe4037586436cb386966b38d81