Do instant indexing with configurable limit on start
Handles the case where a device with a large number of on-device
contacts upgrades from Android S to T where the contacts indexer feature
is enabled. On the first run of contacts indexer service, only the most
recently updated 1000 (configurable limit) contacts are indexed, and the
remaining, if any, CP2 contacts are left for the full update job to
index.

Previous logic to defer handling CP2 changes until the first full update
is run was suboptimal for newly created Android users (work profile,
guest user, etc.)

Bug: 226489369
Test: atest ContactsIndexerTests
Change-Id: I0c3807545f13476c7529e5c41d985f278dd5fa38