Handle AppSearch RESULT_OUT_OF_SPACE error
While indexing documents, we need to consider the document limit of
AppSearch, which currently is 20k. And we are sharing this limit with
other customers of AppSearch in system_server.

So we rely on a full update to trim the Person corpus in AppSearch:
- It will limit the number of contacts we have in AppSearch. Right now
  it is 10,000. Which means we will have at most 10,000 contacts in
  AppSearch after a full update.
- A regular full update will help reduce the number of contacts if we
  are approaching our document limit.
- If the indexer gets a out-of-space error from AppSearch, it will
  immediately schedule a full update, which hopefully can help reduce
  the # of the contacts and free some space.

Based on the stats we get, we can tweak the limit, or reduce the number
of documents based on # of alive documents in case of out-of-space error
so we can make sure the space is freed.

Bug: 222187514
Test: AppSearchServicesTests

Change-Id: I0589d14ba758e35ce150bbbcec2286f4b7b60d42