Handle potential incompatible schmea changes in mainline module rollback
AppSearch will be a mainline module in T. Because the data is not stored
in a snapshottable storage, and a mainline update can be rolled back
after it is installed, we need to handle the rollback gracefully.

When the rollback happens, an old version of Contacts Indexer might be
run on the data written by a newer version. If we try to apply an
incompatible change in AppSearch with forceOverride true, the data
will be wiped in AppSearch after that. We do want to force override
the schema in this case so the right schema can be applied, but need
a way to reindex all the missing contacts again from CP2.

So we need to handle the incompatible schema change even if this is very
1st release of contacts indexer, due to the possibility of the rollback.

In order to detect if an incompatible schema change has been applied, we
can call setSchema twice:

1) Call setSchema with forceOverride "false" first, and if it succeeds, we are good.
2) If the first setSchema fails, we will call setSchema for the second
time, with forceOverride "true". We know the data will be wiped due to the
incompatible schema change, and can reset timestamps and schedule a delta update immediately
to reindex the deleted contacts.

This way, we can handle the incompatible schema change if we need to
upgrade the schemas, or rollback happens and we need to override an
incompatible schema updated before.

Bug: 230338157
Test: ContactsIndexerTests

Change-Id: I727b22b2ca60f40e6345d71e8b3ab21ff4d76be3