Fix race condition in ContactsIndexerUserInstance
ContactsIndexerUserInstance handles CP2 changes by scheduling a task to
perform an incremental update to the Person corpus on a single threaded
executor. This executor is shutdown when a device user is stopped. Since
the ContentObserver.onChange() and SystemService.onUserStopping() calls
happen on different threads, this race condition can lead to a crash.

Add a shared lock to guard access to the executor and check if it is
shutdown before executing a task on it.

Bug: 230643755
Test: atest CtsAppCloningHostTest:com.android.cts.appcloning.AppCloningHostTest#testCreateCloneUserFile --rerun-until-failure 200
Change-Id: I95d0d5edf4e868b3054da0a32670e3256dcc4c78