Fix update job handling.
Since we schedule a separate job for each user, the
ContactsIndexerMaintenanceService JobService might have onStartJob() called
simultaneously for the various users (or while other updates are
happening). In this case, we're being told by JobScheduler to run the
various jobs simultaneously, so we need to keep track of the cancellation
signal for each user update so we stop the appropriate update when
necessary.

Bug: 231148615
Test: Android builds
Change-Id: I7362fd91aa468191cfb068ab3bda6cfba439589e